Random negatives similarity model: generates a model to predict whether two samples are from the same distribution.
A window of discrete samples is gathered from light curve A, and fed through
a 2-layer neural network for a vector v_A. The same thing is done with a
window of discrete samples from light curve B.

We train a model to have a near-1 value for the dot product of v_A and v_B
when A is the same light curve as B, and 0 otherwise. Hopefully this will
give us vectors for reasonably clustering light curves.

Current normalization is dropout, but we should probably add L2 normalization
as well.