Handle windowing differently: select 'before' and 'after' time windows based on the closest point in each band.
Previously, we selected points before and after the selected time
passed, even if that selected time did not correspond to any actual flux
in every band. This approach may make more sense in extracting dflux/dt
features. However, such that we don't select points very far away in one
band (say it's just not sampled for a long time), a bound is introduced
in the WindowFeatures model.